requirements.txt
streamlit
pandas
numpy
openpyxl
matplotlib
scikit-learn
joblib
google-genai
toml